feat: Remove line breaks.
.pat-content-mirror is intended for single-line inputs which need to be
wrapped if the line is getting too long. Line breaks are not allowed but
were possible until now.
Now they are stripped from the input. Instead a single space character
is used. This also works for copy/pasting where a list would not result
in a long concatenated string but with a space as separator in between.

Ref: scrum-2879